LayoutTests/editing/input/cocoa/writing-suggestions-textarea-multiple-lines.html

<!DOCTYPE html> <!-- webkit-test-runner [ useHardwareKeyboardMode=true ] -->
<html>
<head>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>

<style>
textarea {
    font-size: 16px;
    width: 300px;
    height: 300px;
    caret-color: transparent;
}
</style>
</head>
<body>
<textarea id="textarea">good morning.
</textarea>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Verifies that writing suggestions are inserted at the correct offset within the render tree. Requires WebKitTestRunner");

    const editor = document.getElementById("textarea");
    const shadowRoot = internals.shadowRoot(editor);

    shadowRoot.childNodes[0].appendChild(document.createTextNode("My na"));

    editor.focus();
    editor.selectionStart = 20;
    editor.selectionEnd = 20;

    await UIHelper.setInlinePrediction("name is", 2);
    await UIHelper.ensurePresentationUpdate();

    renderedText = internals.elementRenderTreeAsText(editor)
        .split("\n")
        .filter(line => line.includes("text run"))
        .map(line => line.match(/"(.*)"$/)[1])
        .join("");

    shouldBeEqualToString("renderedText", "good morning. My name is");
    finishJSTest();
});
</script>
</body>
</html>
