LayoutTests/editing/input/cocoa/writing-suggestions-textarea-multiple-lines-expected.txt

Verifies that writing suggestions are inserted at the correct offset within the render tree. Requires WebKitTestRunner

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS renderedText is "good morning. My name is"
PASS successfullyParsed is true

TEST COMPLETE
